FCGH-234 Fix
The calculation of maximum days in a week does not limit to the date notice given.
The end of the week  is limited to the earlier of  the dismissal date or pay day, but if the DNG is in the week and before both the Dismissal Date and the Pay week end, it should use the Date Notice Given.